LayoutTests/media/modern-media-controls/skip-back-button/skip-back-button-expected.txt

Testing the SkipBackButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS skipBackButton.element.localName is "button"
PASS skipBackButton.element.classList.contains("icon") is true
PASS skipBackButton.element.classList.contains("skip-back") is true
PASS skipBackButton.iconName is "interval-skip-back"
PASS skipBackButton.element.style.webkitMaskImage.includes("macOS/interval-skip-back@") became true
PASS iOSSkipBackButton.element.style.webkitMaskImage.includes("iOS/interval-skip-back@") became true
PASS successfullyParsed is true

TEST COMPLETE
